feat: refactor report
