+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.1.8+153",
+  "version": "0.1.8+154",
   "private": false,
   "scripts": {
     "dev": "next dev",
