+++ src/components/cash_flow_statement_report_body/cash_flow_statement_report_body_all.tsx
@@ -29,7 +29,7 @@ const CashFlowStatementReportBodyAll = ({ reportId }: ICashFlowStatementReportBo
     success: getReportFinancialSuccess,
     isLoading: getReportFinancialIsLoading,
   } = APIHandler<CashFlowStatementReport>(
-    APIName.REPORT_FINANCIAL_GET_BY_ID,
+    APIName.REPORT_GET_BY_ID,
     {
       params: {
         companyId: selectedCompany?.id,
