+++ src/components/view_financial_section/view_financial_section.tsx
@@ -115,7 +115,7 @@ const ViewFinancialSection = ({
 
   const { data: reportFinancial, isLoading: getReportFinancialIsLoading } =
     APIHandler<FinancialReport>(
-      APIName.REPORT_FINANCIAL_GET_BY_ID,
+      APIName.REPORT_GET_BY_ID,
       {
         params: {
           companyId: selectedCompany?.id,
