+++ src/constants/api_connection.ts
@@ -55,12 +55,9 @@ export enum APIName {
   JOURNAL_LIST = 'JOURNAL_LIST',
   // JOURNAL_UPDATE = 'JOURNAL_UPDATE',
   JOURNAL_DELETE = 'JOURNAL_DELETE',
-  REPORT_LIST_PENDING = 'REPORT_LIST_PENDING',
-  REPORT_LIST_GENERATED = 'REPORT_LIST_GENERATED',
-  REPORT_FINANCIAL_GET_BY_ID = 'REPORT_FINANCIAL_GET_BY_ID',
-  REPORT_ANALYSIS_GET_BY_ID = 'REPORT_ANALYSIS_GET_BY_ID',
-  REPORT_GENERATE_FINANCIAL = 'REPORT_GENERATE_FINANCIAL',
-  REPORT_GENERATE_ANALYSIS = 'REPORT_GENERATE_ANALYSIS',
+  REPORT_LIST = 'REPORT_LIST',
+  REPORT_GET_BY_ID = 'REPORT_GET_BY_ID',
+  REPORT_GENERATE = 'REPORT_GENERATE',
   SESSION_GET = 'SESSION_GET',
   ACCOUNT_LIST = 'ACCOUNT_LIST',
   FILE_UPLOAD = 'FILE_UPLOAD',
@@ -119,12 +116,9 @@ export enum APIPath {
   JOURNAL_LIST = `${apiPrefix}/company/:companyId/journal`,
   // JOURNAL_UPDATE = `${apiPrefix}/company/:companyId/journal/:journalId`,
   JOURNAL_DELETE = `${apiPrefix}/company/:companyId/journal/:journalId`,
-  REPORT_LIST_PENDING = `${apiPrefix}/company/:companyId/report_pending`,
-  REPORT_LIST_GENERATED = `${apiPrefix}/company/:companyId/report_generated`,
-  REPORT_FINANCIAL_GET_BY_ID = `${apiPrefix}/company/:companyId/report_financial/:reportId`,
-  REPORT_ANALYSIS_GET_BY_ID = `${apiPrefix}/company/:companyId/report_analysis/:reportId`,
-  REPORT_GENERATE_FINANCIAL = `${apiPrefix}/company/:companyId/report_financial`,
-  REPORT_GENERATE_ANALYSIS = `${apiPrefix}/company/:companyId/report_analysis`,
+  REPORT_LIST = `${apiPrefix}/company/:companyId/report`,
+  REPORT_GET_BY_ID = `${apiPrefix}/company/:companyId/report/:reportId`,
+  REPORT_GENERATE = `${apiPrefix}/company/:companyId/report`,
   SESSION_GET = `${apiPrefix}/session`,
   ACCOUNT_LIST = `${apiPrefix}/company/:companyId/account`,
   FILE_UPLOAD = `${apiPrefix}/company/:companyId/file`,
@@ -335,36 +329,20 @@ export const APIConfig: Record<IAPIName, IAPIConfig> = {
     method: HttpMethod.PUT,
     path: APIPath.COMPANY_SELECT,
   }),
-  REPORT_LIST_PENDING: createConfig({
-    name: APIName.REPORT_LIST_PENDING,
+  [APIName.REPORT_LIST]: createConfig({
+    name: APIName.REPORT_LIST,
     method: HttpMethod.GET,
-    path: APIPath.REPORT_LIST_PENDING,
+    path: APIPath.REPORT_LIST,
   }),
-  REPORT_LIST_GENERATED: createConfig({
-    name: APIName.REPORT_LIST_GENERATED,
+  [APIName.REPORT_GET_BY_ID]: createConfig({
+    name: APIName.REPORT_GET_BY_ID,
     method: HttpMethod.GET,
-    path: APIPath.REPORT_LIST_GENERATED,
+    path: APIPath.REPORT_GET_BY_ID,
   }),
-
-  [APIName.REPORT_FINANCIAL_GET_BY_ID]: createConfig({
-    name: APIName.REPORT_FINANCIAL_GET_BY_ID,
-    method: HttpMethod.GET,
-    path: APIPath.REPORT_FINANCIAL_GET_BY_ID,
-  }),
-  [APIName.REPORT_ANALYSIS_GET_BY_ID]: createConfig({
-    name: APIName.REPORT_ANALYSIS_GET_BY_ID,
-    method: HttpMethod.GET,
-    path: APIPath.REPORT_ANALYSIS_GET_BY_ID,
-  }),
-  [APIName.REPORT_GENERATE_FINANCIAL]: createConfig({
-    name: APIName.REPORT_GENERATE_FINANCIAL,
+  [APIName.REPORT_GENERATE]: createConfig({
+    name: APIName.REPORT_GENERATE,
     method: HttpMethod.POST,
-    path: APIPath.REPORT_GENERATE_FINANCIAL,
-  }),
-  [APIName.REPORT_GENERATE_ANALYSIS]: createConfig({
-    name: APIName.REPORT_GENERATE_ANALYSIS,
-    method: HttpMethod.POST,
-    path: APIPath.REPORT_GENERATE_ANALYSIS,
+    path: APIPath.REPORT_GENERATE,
   }),
   [APIName.LABOR_COST_CHART]: createConfig({
     name: APIName.LABOR_COST_CHART,
