+++ src/interfaces/api_connection.ts
@@ -38,12 +38,9 @@ export type IAPIName =
   | 'JOURNAL_LIST'
   // | 'JOURNAL_UPDATE'
   | 'JOURNAL_DELETE'
-  | 'REPORT_LIST_PENDING'
-  | 'REPORT_LIST_GENERATED'
-  | 'REPORT_FINANCIAL_GET_BY_ID'
-  | 'REPORT_ANALYSIS_GET_BY_ID'
-  | 'REPORT_GENERATE_FINANCIAL'
-  | 'REPORT_GENERATE_ANALYSIS'
+  | 'REPORT_LIST'
+  | 'REPORT_GET_BY_ID'
+  | 'REPORT_GENERATE'
   | 'SESSION_GET'
   | 'ACCOUNT_LIST'
   | 'FILE_UPLOAD'
